pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "lung_alignment"
version = "2.1.0"
description = "COVID-19 Detection via Anatomical Landmarks and Geometric Normalization"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Proyecto de Tesis"}
]
keywords = [
    "covid-19",
    "deep-learning",
    "computer-vision",
    "chest-x-ray",
    "landmarks",
    "geometric-normalization",
    "piecewise-affine-warping"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=2.0.0",
    "scipy>=1.10.0",
    "pandas>=2.0.0",
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "tqdm>=4.65.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
lung-alignment = "src_v2.cli:app"

# URLs comentadas hasta tener repositorio publico
# [project.urls]
# "Homepage" = "https://github.com/<usuario>/prediccion_warping_clasificacion"
# "Bug Tracker" = "https://github.com/<usuario>/prediccion_warping_clasificacion/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src_v2*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
env = [
    "FORCE_NUM_WORKERS_ZERO=1",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

[tool.coverage.run]
source = ["src_v2"]
omit = ["tests/*", "scripts/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
